Move snap macros to expconf module
